requirements_windows.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
accelerate=0.22.0=pypi_0
blas=1.0=mkl
ca-certificates=2023.01.10=haa95532_0
certifi=2022.12.7=py38haa95532_0
charset-normalizer=3.2.0=pypi_0
colorama=0.4.6=pypi_0
cuda-cccl=12.2.140=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
diffusers=0.20.2=pypi_0
filelock=3.9.0=py38haa95532_0
flit-core=3.6.0=pyhd3eb1b0_0
fsspec=2023.9.0=pypi_0
huggingface-hub=0.17.1=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
intel-openmp=2022.1.0=h59b6b97_3788
jinja2=3.1.2=py38haa95532_0
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.3.141=0
libcurand-dev=10.3.3.141=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libffi=3.4.2=hd77b12b_6
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libuv=1.44.2=h2bbff1b_0
markupsafe=2.1.1=py38h2bbff1b_0
mkl=2022.1.0=h858461b_193
mpmath=1.2.1=py38haa95532_0
networkx=2.8.4=py38haa95532_0
numpy=1.24.4=pypi_0
opencv-python=4.8.0.76=pypi_0
openssl=1.1.1s=h2bbff1b_0
packaging=23.1=pypi_0
pillow=10.0.0=pypi_0
pip=22.3.1=py38haa95532_0
psutil=5.9.5=pypi_0
python=3.8.16=h6244533_2
pytorch=2.0.1=py3.8_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=pypi_0
regex=2023.8.8=pypi_0
requests=2.31.0=pypi_0
safetensors=0.3.3=pypi_0
setuptools=65.6.3=py38haa95532_0
sqlite=3.40.1=h2bbff1b_0
sympy=1.11.1=py38haa95532_0
tokenizers=0.13.3=pypi_0
tqdm=4.66.1=pypi_0
transformers=4.33.1=pypi_0
typing_extensions=4.4.0=py38haa95532_0
urllib3=2.0.4=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
xformers=0.0.21=pypi_0
zipp=3.16.2=pypi_0
